% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unemployment.r
\docType{data}
\name{unemployment}
\alias{unemployment}
\title{Labor Force Data by County, 2015 Annual Averages}
\format{A data frame with 3219 rows and 9 variables}
\source{
\url{https://www.bls.gov/lau/laucnty15.txt}
}
\usage{
data(unemployment)
}
\description{
Labor force data provided by the US Bureau of Labor and Statistics.
From their website (at \url{https://www.bls.gov/lau/}):

The Local Area Unemployment Statistics (LAUS) program
produces monthly and annual employment, unemployment, and labor
force data for Census regions and divisions, States, counties,
metropolitan areas, and many cities, by place of residence.
}
\details{
The dataset contains the following variables:

\itemize{
  \item LAUS Code
  \item State FIPS Code
  \item County FIPS Code
  \item County Name/State Abbr.
  \item Year
  \item Labor Force
  \item Employed
  \item Unemployment Level
  \item Unemployment Rate
}
}
\keyword{datasets}
